+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/HostClassCache.java
@@ -40,6 +40,7 @@
  */
 package com.oracle.truffle.polyglot;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
@@ -49,6 +50,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import org.graalvm.collections.EconomicSet;
 import org.graalvm.polyglot.HostAccess;
 import org.graalvm.polyglot.impl.AbstractPolyglotImpl;
 import org.graalvm.polyglot.impl.AbstractPolyglotImpl.APIAccess;
@@ -60,7 +62,7 @@ final class HostClassCache {
     static final PolyglotTargetMapping[] EMPTY_MAPPINGS = new PolyglotTargetMapping[0];
 
     private final APIAccess apiAccess;
-    private final HostAccess hostAccess;
+    final HostAccess hostAccess;
     private final boolean arrayAccess;
     private final boolean listAccess;
     private final Map<Class<?>, Object> targetMappings;
@@ -175,14 +177,17 @@ final class HostClassCache {
         return descs.get(clazz);
     }
 
+    @TruffleBoundary
     boolean allowsAccess(Method m) {
         return apiAccess.allowsAccess(hostAccess, m);
     }
 
+    @TruffleBoundary
     boolean allowsAccess(Constructor<?> m) {
         return apiAccess.allowsAccess(hostAccess, m);
     }
 
+    @TruffleBoundary
     boolean allowsAccess(Field f) {
         return apiAccess.allowsAccess(hostAccess, f);
     }
@@ -193,6 +198,10 @@ final class HostClassCache {
 
     boolean isListAccess() {
         return listAccess;
+    }
+
+    EconomicSet<Class<? extends Annotation>> getExportAnnotations() {
+        return apiAccess.getExportAnnotations(hostAccess);
     }
 
 }